Separate measurements in different files.
